pyproject.toml
[tool.poetry]
name = "natalia"
version = "2.0.0"
description = "discord bot"
authors = ["konafx <afxlieonon@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.4"
"discord.py" = "^1.5.0"
more_itertools = "^8.2.0"
kanjize = "^0.1.0"
emoji = "^0.6.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
pytest = "^5.4.2"
pytest-asyncio = "^0.12.0"

[tool.poetry.scripts]
natalia = "natalia:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
